+++ rustboy-gb/src/interconnect/cartridge/cartridge_info.rs
